refactor: рефакторит код проекта
* Добавляет алиас для путей
* Изменяет файловую структуру
* Определяет порядок импортов: внешние либы, встроенные либы, внешние
  директории, текущая директория
