+++ src/cli/app/lib/command-parser.interface.ts
@@ -1,0 +1,5 @@
+import { ParsedCommands } from "./parsed-commands.type.js";
+
+export interface ICommandParser {
+  parse(argv: string[]): ParsedCommands;
+}
