+++ src/cli/app/lib/parsed-commands.type.ts
@@ -1,0 +1,1 @@
+export type ParsedCommands = Record<string, string[]>;
