+++ src/cli/command/command.interface.ts
