+++ src/cli/command/lib/tsv/file-reader/index.ts
@@ -1,0 +1,1 @@
+export { TSVFileReader } from "./tsv-file-reader.js";
