+++ src/cli/command/lib/tsv/file-reader/tsv-file-reader.interface.ts
@@ -1,3 +1,3 @@
-export interface IFileReader {
+export interface ITSVFileReader {
   read(filePath?: string): void;
 }
