--- src/shared/lib/file-reader/index.ts
@@ -1,1 +1,0 @@
-export { TSVFileReader } from './tsv/tsv-file-reader.js';
