+++ tsconfig.json
@@ -25,16 +25,21 @@
     ],
     "lib": [
       "ESNext"
-    ]
+    ],
+    "baseUrl": ".",
+    "paths": {
+      "@/*": [
+        "src/*",
+        "mocks/*",
+      ]
+    },
   },
   "include": [
-    "src/**/*.ts"
+    "src/**/*.ts",
+    "mocks/**/*.ts"
   ],
   "exclude": [
     "node_modules",
     "**/*.test.ts"
-  ],
-  "ts-node": {
-    "esm": true
-  }
+  ]
 }
